fixat så det är float på de flesta ställena
Kanske fortfarande inte funkar, men det får vara så då

+++ src/circularBuffer.h
@@ -64,7 +64,7 @@ int getSize(struct circularBuffer* bufferPtr);
 
 int getSum(struct circularBuffer* bufferPtr);
 
-int getAverage(struct circularBuffer* bufferPtr);
+float getAverage(struct circularBuffer* bufferPtr);
 
 #endif
